Monday fixes

+++ app/controllers/validations_controller.rb
@@ -17,7 +17,7 @@ class ValidationsController < ApplicationController
   filter_access_to :edit, :show, :attribute_check=>true
   filter_access_to :admin_manage, :attribute_check=>true
   filter_access_to :send_to_ca, require: :sysadmin_manage
-  filter_access_to :get_asynch_domains, :remove_domains, :get_email_addresses, :require=>:new
+  filter_access_to :get_asynch_domains, :remove_domains, :get_email_addresses, :require=>:ajax
   in_place_edit_for :validation_history, :notes
 
   def search
@@ -75,52 +75,60 @@ class ValidationsController < ApplicationController
   end
 
   def remove_domains
-    domain_name_arry = params['domain_names'].split(',')
-    # order_number = CertificateOrder.find_by_ref(params['certificate_order_id']).external_order_number
-    certificate_order = (current_user.is_system_admins? ? CertificateOrder :
-         current_user.ssl_account.certificate_orders).find_by_ref(params[:certificate_order_id])
     result_obj = {}
+    if current_user
+      domain_name_arry = params['domain_names'].split(',')
+      # order_number = CertificateOrder.find_by_ref(params['certificate_order_id']).external_order_number
+      certificate_order = (current_user.is_system_admins? ? CertificateOrder :
+                               current_user.ssl_account.certificate_orders).find_by_ref(params[:certificate_order_id])
 
-    domain_name_arry.each do |domain_name|
-      cn_obj = certificate_order.certificate_content.certificate_names.find_by_name(domain_name)
-      next unless cn_obj
+      domain_name_arry.each do |domain_name|
+        cn_obj = certificate_order.certificate_content.certificate_names.find_by_name(domain_name)
+        next unless cn_obj
 
-      res = ComodoApi.auto_remove_domain(domain_name: cn_obj, order_number: certificate_order.external_order_number)
+        res = ComodoApi.auto_remove_domain(domain_name: cn_obj, order_number: certificate_order.external_order_number)
 
-      error_code = -1
-      error_message = ''
+        error_code = -1
+        error_message = ''
 
-      if res.index('errorCode') && res.index('errorMessage')
-        error_code = res.split('&')[0].split('=')[1].to_i
-        error_message = res.split('&')[1].split('=')[1]
-      elsif res.index('errorCode') && !res.index('errorMessage')
-        error_code = 0
-      else
-        error_message = res
+        if res.index('errorCode') && res.index('errorMessage')
+          error_code = res.split('&')[0].split('=')[1].to_i
+          error_message = res.split('&')[1].split('=')[1]
+        elsif res.index('errorCode') && !res.index('errorMessage')
+          error_code = 0
+        else
+          error_message = res
+        end
+
+        if error_code.zero?
+          cn_obj.destroy
+        else
+          result_obj[domain_name] = error_message.gsub("+", " ").gsub("%27", "'").gsub("%21", "!")
+        end
       end
-
-      if error_code.zero?
-        cn_obj.destroy
-      else
-        result_obj[domain_name] = error_message.gsub("+", " ").gsub("%27", "'").gsub("%21", "!")
-      end
+    else
+      result_obj['no-user'] = "true"
     end
 
     render :json => result_obj
   end
 
   def get_email_addresses
-    addresses = params['total_domains'].to_i > Validation::COMODO_EMAIL_LOOKUP_THRESHHOLD ?
-                    DomainControlValidation.email_address_choices(params['domain_name']) :
-                    ComodoApi.domain_control_email_choices(params['domain_name']).email_address_choices
-    addresses.delete("none")
-
     returnObj = {}
-    returnObj['caa_check'] = CaaCheck.pass?(params[:certificate_order_id], params['domain_name']) ? 'passed' : 'failed'
-    returnObj['new_emails'] = {}
+    if current_user
+      addresses = params['total_domains'].to_i > Validation::COMODO_EMAIL_LOOKUP_THRESHHOLD ?
+                      DomainControlValidation.email_address_choices(params['domain_name']) :
+                      ComodoApi.domain_control_email_choices(params['domain_name']).email_address_choices
+      addresses.delete("none")
 
-    addresses.each do |addr|
-      returnObj['new_emails'][addr] = addr
+      returnObj['caa_check'] = CaaCheck.pass?(params[:certificate_order_id], params['domain_name']) ? 'passed' : 'failed'
+      returnObj['new_emails'] = {}
+
+      addresses.each do |addr|
+        returnObj['new_emails'][addr] = addr
+      end
+    else
+      returnObj['no-user'] = "true"
     end
 
     render :json => returnObj
